Develop -> 가연 250624

+++ src/main/java/com/team/arium/domain/Cnsl_PreEvalId.java
@@ -8,5 +8,5 @@ import java.io.Serializable;
 @AllArgsConstructor
 public class Cnsl_PreEvalId implements Serializable {
     private String preEvalId; 
-    private Long preQstId;
+    private Integer preQstId;
 }
